requirements.txt
fastapi
uvicorn
motor==2.5.1
pymongo[srv]
python-dotenv